+++ src/main/java/com/woowacourse/sunbook/application/user/LoginService.java
@@ -1,9 +1,9 @@
-package com.woowacourse.sunbook.application;
+package com.woowacourse.sunbook.application.user;
 
-import com.woowacourse.sunbook.application.dto.user.UserRequestDto;
-import com.woowacourse.sunbook.application.dto.user.UserResponseDto;
 import com.woowacourse.sunbook.application.exception.DuplicateEmailException;
 import com.woowacourse.sunbook.application.exception.LoginException;
+import com.woowacourse.sunbook.application.user.dto.UserRequestDto;
+import com.woowacourse.sunbook.application.user.dto.UserResponseDto;
 import com.woowacourse.sunbook.domain.user.User;
 import com.woowacourse.sunbook.domain.user.UserRepository;
 import org.modelmapper.ModelMapper;
