+++ src/main/java/com/woowacourse/sunbook/support/ExceptionAdvice.java
@@ -1,10 +1,9 @@
-package com.woowacourse.sunbook.presentation;
+package com.woowacourse.sunbook.support;
 
 import com.fasterxml.jackson.databind.exc.InvalidDefinitionException;
 import com.woowacourse.sunbook.application.exception.DuplicateEmailException;
 import com.woowacourse.sunbook.application.exception.LoginException;
 import com.woowacourse.sunbook.domain.validation.exception.InvalidValueException;
-import com.woowacourse.sunbook.support.ErrorMessage;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.converter.HttpMessageConversionException;
@@ -12,7 +11,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 @RestControllerAdvice
-public class ExceptionController {
+public class ExceptionAdvice {
 
     @ExceptionHandler({DuplicateEmailException.class, LoginException.class, InvalidValueException.class})
     public ResponseEntity<ErrorMessage> responseExceptionMessage(RuntimeException exception) {
