+++ src/test/java/com/woowacourse/sunbook/application/ArticleServiceTest.java
@@ -1,6 +1,7 @@
 package com.woowacourse.sunbook.application;
 
-import com.woowacourse.sunbook.application.dto.article.ArticleResponseDto;
+import com.woowacourse.sunbook.application.article.ArticleService;
+import com.woowacourse.sunbook.application.article.dto.ArticleResponseDto;
 import com.woowacourse.sunbook.domain.article.Article;
 import com.woowacourse.sunbook.domain.article.ArticleFeature;
 import com.woowacourse.sunbook.domain.article.ArticleRepository;
