+++ src/test/java/com/woowacourse/sunbook/application/LoginServiceTest.java
@@ -1,9 +1,10 @@
 package com.woowacourse.sunbook.application;
 
-import com.woowacourse.sunbook.application.dto.user.UserRequestDto;
-import com.woowacourse.sunbook.application.dto.user.UserResponseDto;
 import com.woowacourse.sunbook.application.exception.DuplicateEmailException;
 import com.woowacourse.sunbook.application.exception.LoginException;
+import com.woowacourse.sunbook.application.user.LoginService;
+import com.woowacourse.sunbook.application.user.dto.UserRequestDto;
+import com.woowacourse.sunbook.application.user.dto.UserResponseDto;
 import com.woowacourse.sunbook.domain.user.User;
 import com.woowacourse.sunbook.domain.user.UserEmail;
 import com.woowacourse.sunbook.domain.user.UserPassword;
