+++ src/test/java/com/woowacourse/sunbook/presentation/ArticleApiControllerTest.java
@@ -8,7 +8,7 @@ import org.springframework.http.HttpStatus;
 
 import static org.hamcrest.Matchers.hasItem;
 
-class ArticleControllerTest extends TestTemplate {
+class ArticleApiControllerTest extends TestTemplate {
     private static final String CONTENTS = "SunBook Contents";
     private static final String IMAGE_URL = "https://file.namu.moe/file/105db7e730e1402c09dcf2b281232df017f0966ba63375176cb0886869b81bf206145de5a7a149a987d6aae2d5230afaae4ca2bf0b418241957942ad4f4a08c8";
     private static final String VIDEO_URL = "https://youtu.be/mw5VIEIvuMI";
